requirements.txt
beautifulsoup4==4.12.3
pillow==10.3.0
pytesseract==0.3.10
regex==2024.5.15
requests==2.32.3
torch==2.3.1